Keep original pivot points
Looks like even if we keep the pivot point and import the single mesh back into Blender it does keep it, but Stride still pivots from origin location (0,0,0) -
extremely annoying.

Found Issue: From testing, it looks like the only way to allow Stride to pivot
correctly is to reset the object transform to be at the pivot before export.
This makes it hard to pre-setup things like levels because if a mesh is away
from origin (car windows, whatever), it will use world origin as pivot, not the
actual mesh's pivot.

Workaround: When you reset your mesh transform, putting it at world origin at its pivot point, it sort of fixes it.